Moved network monitor related files to separate package
