+++ validator-api/src/main.rs
@@ -4,44 +4,26 @@
 #[macro_use]
 extern crate rocket;
 
+use crate::network_monitor::new_monitor_runnables;
+use crate::network_monitor::tested_network::good_topology::parse_topology_file;
+use crate::node_status_api::storage::NodeStatusStorage;
+use anyhow::Result;
+use cache::ValidatorCache;
+use clap::{App, Arg, ArgMatches};
+use log::info;
+use mixnet_contract::{GatewayBond, MixNodeBond};
 use rocket::http::Method;
 use rocket::serde::json::Json;
 use rocket::State;
 use rocket_cors::{AllowedHeaders, AllowedOrigins};
-
-use crate::monitor::preparer::PacketPreparer;
-use crate::monitor::processor::{
-    ReceivedProcessor, ReceivedProcessorReceiver, ReceivedProcessorSender,
-};
-use crate::monitor::receiver::{
-    GatewayClientUpdateReceiver, GatewayClientUpdateSender, PacketReceiver,
-};
-use crate::monitor::sender::PacketSender;
-use crate::monitor::summary_producer::SummaryProducer;
-use crate::node_status_api::storage::NodeStatusStorage;
-use crate::tested_network::good_topology::parse_topology_file;
-use crate::tested_network::TestedNetwork;
-use anyhow::Result;
-use cache::ValidatorCache;
-use clap::{App, Arg, ArgMatches};
-use crypto::asymmetric::{encryption, identity};
-use futures::channel::mpsc;
-use log::info;
-use mixnet_contract::{GatewayBond, MixNodeBond};
-use nymsphinx::addressing::clients::Recipient;
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::RwLock;
 use tokio::time;
-use topology::NymTopology;
 
 mod cache;
-mod chunker;
-pub(crate) mod gateways_reader;
-mod monitor;
+mod network_monitor;
 mod node_status_api;
-mod test_packet;
-mod tested_network;
 
 const V4_TOPOLOGY_ARG: &str = "v4-topology-filepath";
 const V6_TOPOLOGY_ARG: &str = "v6-topology-filepath";
@@ -133,74 +115,6 @@ async fn wait_for_interrupt() {
     println!("Received SIGINT - the network monitor will terminate now");
 }
 
-fn new_packet_preparer(
-    validator_client: validator_client::Client,
-    tested_network: TestedNetwork,
-    test_mixnode_sender: Recipient,
-    self_public_identity: identity::PublicKey,
-    self_public_encryption: encryption::PublicKey,
-) -> PacketPreparer {
-    PacketPreparer::new(
-        validator_client,
-        tested_network,
-        test_mixnode_sender,
-        self_public_identity,
-        self_public_encryption,
-    )
-}
-
-fn new_packet_sender(
-    gateways_status_updater: GatewayClientUpdateSender,
-    local_identity: Arc<identity::KeyPair>,
-    max_sending_rate: usize,
-) -> PacketSender {
-    PacketSender::new(
-        gateways_status_updater,
-        local_identity,
-        GATEWAY_RESPONSE_TIMEOUT,
-        MAX_CONCURRENT_GATEWAY_CLIENTS,
-        max_sending_rate,
-    )
-}
-
-fn new_received_processor(
-    packets_receiver: ReceivedProcessorReceiver,
-    client_encryption_keypair: Arc<encryption::KeyPair>,
-) -> ReceivedProcessor {
-    ReceivedProcessor::new(packets_receiver, client_encryption_keypair)
-}
-
-fn new_summary_producer(detailed_report: bool) -> SummaryProducer {
-    // right now always print the basic report. If we feel like we need to change it, it can
-    // be easily adjusted by adding some flag or something
-    let summary_producer = SummaryProducer::default().with_report();
-    if detailed_report {
-        summary_producer.with_detailed_report()
-    } else {
-        summary_producer
-    }
-}
-
-fn new_packet_receiver(
-    gateways_status_updater: GatewayClientUpdateReceiver,
-    processor_packets_sender: ReceivedProcessorSender,
-) -> PacketReceiver {
-    PacketReceiver::new(gateways_status_updater, processor_packets_sender)
-}
-
-fn new_validator_client(
-    validator_rest_uris: Vec<String>,
-    mixnet_contract: &str,
-) -> validator_client::Client {
-    let config = validator_client::Config::new(validator_rest_uris, mixnet_contract);
-    validator_client::Client::new(config)
-}
-
-fn new_node_status_api_client<S: Into<String>>(base_url: S) -> node_status_api::Client {
-    let config = node_status_api::Config::new(base_url);
-    node_status_api::Client::new(config)
-}
-
 fn setup_logging() {
     let mut log_builder = pretty_env_logger::formatted_timed_builder();
     if let Ok(s) = ::std::env::var("RUST_LOG") {
@@ -222,49 +136,6 @@ fn setup_logging() {
         .init();
 }
 
-fn check_if_up_to_date(v4_topology: &NymTopology, v6_topology: &NymTopology) {
-    let monitor_version = env!("CARGO_PKG_VERSION");
-    for (_, layer_mixes) in v4_topology.mixes().iter() {
-        for mix in layer_mixes.iter() {
-            if !version_checker::is_minor_version_compatible(monitor_version, &*mix.version) {
-                panic!(
-                    "Our good topology is not compatible with monitor! Mix runs {}, we have {}",
-                    mix.version, monitor_version
-                )
-            }
-        }
-    }
-
-    for gateway in v4_topology.gateways().iter() {
-        if !version_checker::is_minor_version_compatible(monitor_version, &*gateway.version) {
-            panic!(
-                "Our good topology is not compatible with monitor! Gateway runs {}, we have {}",
-                gateway.version, monitor_version
-            )
-        }
-    }
-
-    for (_, layer_mixes) in v6_topology.mixes().iter() {
-        for mix in layer_mixes.iter() {
-            if !version_checker::is_minor_version_compatible(monitor_version, &*mix.version) {
-                panic!(
-                    "Our good topology is not compatible with monitor! Mix runs {}, we have {}",
-                    mix.version, monitor_version
-                )
-            }
-        }
-    }
-
-    for gateway in v6_topology.gateways().iter() {
-        if !version_checker::is_minor_version_compatible(monitor_version, &*gateway.version) {
-            panic!(
-                "Our good topology is not compatible with monitor! Gateway runs {}, we have {}",
-                gateway.version, monitor_version
-            )
-        }
-    }
-}
-
 #[get("/mixnodes")]
 async fn get_mixnodes(cache: &State<Arc<RwLock<ValidatorCache>>>) -> Json<Vec<MixNodeBond>> {
     let cache = cache.read().await;
@@ -280,6 +151,7 @@ async fn get_gateways(cache: &State<Arc<RwLock<ValidatorCache>>>) -> Json<Vec<Ga
 #[tokio::main]
 async fn main() -> Result<()> {
     info!("Network monitor starting...");
+
     let matches = parse_args();
     let v4_topology_path = matches.value_of(V4_TOPOLOGY_ARG).unwrap();
     let v6_topology_path = matches.value_of(V6_TOPOLOGY_ARG).unwrap();
@@ -317,7 +189,7 @@ async fn main() -> Result<()> {
         .map(|v| v.parse().unwrap())
         .unwrap_or_else(|| DEFAULT_CACHE_INTERVAL_ARG);
 
-    check_if_up_to_date(&v4_topology, &v6_topology);
+    network_monitor::check_if_up_to_date(&v4_topology, &v6_topology);
     setup_logging();
 
     info!("* validator servers: {:?}", validators_rest_uris);
@@ -326,70 +198,19 @@ async fn main() -> Result<()> {
     info!("* detailed report printing: {}", detailed_report);
     info!("* gateway sending rate: {} packets/s", sending_rate);
 
-    // TODO: in the future I guess this should somehow change to distribute the load
-    let tested_mix_gateway = v4_topology.gateways()[0].clone();
-    info!(
-        "* gateway for testing mixnodes: {}",
-        tested_mix_gateway.identity_key.to_base58_string()
-    );
-
-    // TODO: those keys change constant throughout the whole execution of the monitor.
-    // and on top of that, they are used with ALL the gateways -> presumably this should change
-    // in the future
-    let mut rng = rand::rngs::OsRng;
-
-    let identity_keypair = Arc::new(identity::KeyPair::new(&mut rng));
-    let encryption_keypair = Arc::new(encryption::KeyPair::new(&mut rng));
-
-    let test_mixnode_sender = Recipient::new(
-        *identity_keypair.public_key(),
-        *encryption_keypair.public_key(),
-        tested_mix_gateway.identity_key,
-    );
-
-    let tested_network = TestedNetwork::new_good(v4_topology, v6_topology);
-    let validator_client = new_validator_client(validators_rest_uris.clone(), &mixnet_contract);
-    let node_status_api_client = new_node_status_api_client(node_status_api_uri);
-
-    let (gateway_status_update_sender, gateway_status_update_receiver) = mpsc::unbounded();
-    let (received_processor_sender_channel, received_processor_receiver_channel) =
-        mpsc::unbounded();
-
-    let packet_preparer = new_packet_preparer(
-        validator_client,
-        tested_network.clone(),
-        test_mixnode_sender,
-        *identity_keypair.public_key(),
-        *encryption_keypair.public_key(),
-    );
-
-    let packet_sender = new_packet_sender(
-        gateway_status_update_sender,
-        Arc::clone(&identity_keypair),
+    let network_monitor_runnables = new_monitor_runnables(
+        validators_rest_uris.clone(),
+        &mixnet_contract,
+        &node_status_api_uri,
+        v4_topology,
+        v6_topology,
         sending_rate,
-    );
-    let received_processor = new_received_processor(
-        received_processor_receiver_channel,
-        Arc::clone(&encryption_keypair),
-    );
-    let summary_producer = new_summary_producer(detailed_report);
-    let mut packet_receiver = new_packet_receiver(
-        gateway_status_update_receiver,
-        received_processor_sender_channel,
-    );
-
-    let mut monitor = monitor::Monitor::new(
-        packet_preparer,
-        packet_sender,
-        received_processor,
-        summary_producer,
-        node_status_api_client,
-        tested_network,
+        detailed_report,
     );
 
     let mixnode_cache = Arc::new(RwLock::new(ValidatorCache::init(
-        validators_rest_uris,
-        mixnet_contract,
+        vec!["validators_rest_uris".to_string()],
+        "mixnet_contract".to_string(),
     )));
 
     let write_mixnode_cache = Arc::clone(&mixnode_cache);
@@ -408,7 +229,7 @@ async fn main() -> Result<()> {
         }
     });
 
-    tokio::spawn(async move { packet_receiver.run().await });
+    network_monitor_runnables.spawn_tasks();
 
     let node_status_storage = NodeStatusStorage::new();
