+++ validator-api/src/network_monitor/gateways_reader.rs
