+++ validator-api/src/network_monitor/monitor/receiver.rs
@@ -1,8 +1,19 @@
-// Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
+// Copyright 2021 Nym Technologies SA
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-use crate::gateways_reader::{GatewayChannel, GatewayMessages, GatewaysReader};
-use crate::monitor::processor::ReceivedProcessorSender;
+use crate::network_monitor::gateways_reader::{GatewayChannel, GatewayMessages, GatewaysReader};
+use crate::network_monitor::monitor::processor::ReceivedProcessorSender;
 use crypto::asymmetric::identity;
 use futures::channel::mpsc;
 use futures::StreamExt;
