+++ validator-api/src/network_monitor/monitor/summary_producer.rs
@@ -1,11 +1,22 @@
-// Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
+// Copyright 2021 Nym Technologies SA
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-use crate::monitor::preparer::{InvalidNode, TestedNode};
+use crate::network_monitor::monitor::preparer::{InvalidNode, TestedNode};
+use crate::network_monitor::test_packet::{NodeType, TestPacket};
 use crate::node_status_api::models::{
     BatchGatewayStatus, BatchMixStatus, GatewayStatus, MixStatus,
 };
-use crate::test_packet::{NodeType, TestPacket};
 use crate::PENALISE_OUTDATED;
 use std::collections::HashMap;
