+++ validator-api/src/network_monitor/test_packet.rs
@@ -1,7 +1,18 @@
-// Copyright 2020 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
+// Copyright 2021 Nym Technologies SA
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-use crate::monitor::preparer::TestedNode;
+use crate::network_monitor::monitor::preparer::TestedNode;
 use crypto::asymmetric::identity;
 use std::convert::{TryFrom, TryInto};
 use std::fmt::{self, Display, Formatter};
